added new method of intergration & changed the default one
