+++ project/include/whfast.hpp
@@ -1,0 +1,41 @@
+/**
+ * @file whfast.hpp
+ * @brief Wisdom-Holman fast — симплектический интегратор для систем
+ *        с одним доминирующим телом (Solar-like).
+ *
+ * Гамильтониан расщепляется в democratic-heliocentric координатах на:
+ *  - H_K — Кеплер-задача каждой планеты вокруг центрального тела (точно).
+ *  - H_I — попарные планет-планетные возмущения (kick).
+ *  - H_J — «Sun jump» (drift, обусловленный отсутствием тяжёлой звезды
+ *           в Кеплер-части).
+ *
+ * Один шаг WHFast 2-го порядка:
+ *     I(h/2) ∘ J(h/2) ∘ K(h) ∘ J(h/2) ∘ I(h/2)
+ *
+ * Кеплер решается через универсальные переменные χ и функции Стумпффа.
+ *
+ * @par Применимость
+ *  - Solar-like системы (одно сильно доминирующее тело + N≪1000 спутников).
+ *  - НЕ для globular cluster / однородных распределений.
+ *
+ * @par Точность
+ *  При шаге h ≈ T_min/30 ошибка позиции на оборот ~10⁻¹⁰…10⁻¹², долговременной
+ *  фазовой расходимости нет (симплектичность сохраняется при фиксированном h).
+ */
+
+#pragma once
+#include <vector>
+#include "object.hpp"
+
+/**
+ * @brief Один шаг симуляции методом WHFast (2-й порядок).
+ *
+ * @param objs       Тела (модифицируется in-place).
+ * @param dt         Фиксированный шаг времени [год].
+ * @param pause      Если true — шаг пропускается.
+ * @param forceSync  Игнорируется (CPU-версия).
+ */
+void simulationStepWHFastCPU(std::vector<Object>& objs,
+                             double dt,
+                             bool pause,
+                             bool forceSync = false);
