+++ verifications_tools/system_past.txt
@@ -1,0 +1,11 @@
+# Солнечная система на 2025-03-16
+# Format: Name Mass Density Px Py Pz Vx Vy Vz R G B
+Sun 1.98847e+30 1410 0.000 0.000 0.000 0.000 0.000 0.000 1.0 0.9 0.1
+Mercury 3.3011e+23 5429 -41593525130.801 29474332812.203 6223697230.637 -38117.770 -37728.115 413.001 0.5 0.5 0.5
+Venus 4.8675e+24 5243 -106098222512.141 17015774307.355 6355526511.411 -5731.317 -34734.958 -146.406 0.9 0.8 0.6
+Earth 5.9723e+24 5514 -148298929211.249 12191647695.135 299926.849 -2927.189 -29788.415 2.783 0.2 0.5 1.0
+Mars 6.4171e+23 3934 -196233299090.445 152223046136.763 8002180794.699 -13940.062 -17076.811 -16.018 1.0 0.3 0.1
+Jupiter 1.8982e+27 1326 74418777913.258 760412510532.878 -4823704350.198 -13167.091 1888.512 286.781 0.8 0.7 0.6
+Saturn 5.6834e+26 687 1421915176470.061 -203147059196.341 -53062850989.112 822.173 9546.656 -199.116 0.9 0.8 0.5
+Uranus 8.681e+25 1270 1624658161643.367 2430165998828.439 -12040441963.622 -5724.377 3471.741 86.891 0.6 0.8 0.9
+Neptune 1.0241e+26 1638 4470271330917.463 -59886083007.379 -101780412507.540 25.487 5472.512 -112.729 0.2 0.3 0.8
